All grants

Returns a list of data frames with details of all grants from funders returned by tsg_available().

tsg_all_grants(verbose = TRUE, timeout = 30, retries = 0, correct_names = TRUE)

Arguments

verbose

If TRUE, prints console messages on data retrieval progress. Defaults to TRUE.

timeout

The maximum request time, in seconds. If data is not returned in this time, a value of NA is returned for that dataset. Defaults to 30 seconds.

retries

The number of retries to make if a request is not successful. Defaults to 0.

correct_names

If TRUE, corrects known mistakes in column names, such as spelling mistakes. Defaults to TRUE.

Value

A list of tibbles with grant data.

Details

Due to the structure of the 360 Giving data, the package will make multiple attempts to request data. This can cause issues with servers automatically blocking requests. You can use tsg_missing() to identify missing sets of grant data.

Examples

if (FALSE) {
all_grants <- tsg_all_grants()
}