Core Data

Given a list returned by [tsg_all_grants()] or [tsg_search_grants()], creates a tibble with the core variables required by the [360Giving](https://standard.threesixtygiving.org/en/latest/) open standard, as well as the publisher prefix, which is useful for data processing, and the licence the data was published under.

tsg_core_data(x, verbose = TRUE)

Arguments

x
A list of tibble with grant data returned by [tsg_all_grants()].

verbose
If `TRUE`, prints console messages.

Value

A tibble with core variables

Examples

# \donttest{
grants <- tsg_all_grants()

#> Downloading 1 of 195

#> Downloading 2 of 195

#> Downloading 3 of 195

#> Downloading 4 of 195

#> Downloading 5 of 195

#> New names:
#> * Identifier -> Identifier...1
#> * Identifier -> Identifier...2

#> Downloading 6 of 195

#> Downloading 7 of 195

#> Downloading 8 of 195

#> Downloading 9 of 195

#> Downloading 10 of 195

#> Downloading 11 of 195

#> Downloading 12 of 195

#> Downloading 13 of 195

#> New names:
#> * `` -> ...17

#> Downloading 14 of 195

#> New names:
#> * `` -> ...17

#> Downloading 15 of 195

#> Downloading 16 of 195

#> Downloading 17 of 195

#> Downloading 18 of 195

#> Downloading 19 of 195

#> Downloading 20 of 195

#> Downloading 21 of 195

#> Downloading 22 of 195

#> Downloading 23 of 195

#> Downloading 24 of 195

#> New names:
#> * `` -> ...2

#> Downloading 25 of 195

#> Downloading 26 of 195

#> Downloading 27 of 195

#> Downloading 28 of 195

#> Downloading 29 of 195

#> Downloading 30 of 195

#> Downloading 31 of 195

#> Downloading 32 of 195

#> Downloading 33 of 195

#> Downloading 34 of 195

#> Downloading 35 of 195

#> Downloading 36 of 195

#> Downloading 37 of 195

#> Downloading 38 of 195

#> Downloading 39 of 195

#> Downloading 40 of 195

#> Downloading 41 of 195

#> Downloading 42 of 195

#> Downloading 43 of 195

#> Downloading 44 of 195

#> Downloading 45 of 195

#> Downloading 46 of 195

#> Downloading 47 of 195

#> Downloading 48 of 195

#> Request failed [403]. Retrying in 1 seconds...

#> Request failed [403]. Retrying in 1.6 seconds...

#> Request failed: Client error: (403) Forbidden

#> Downloading 49 of 195

#> Downloading 50 of 195

#> Downloading 51 of 195

#> Downloading 52 of 195

#> Downloading 53 of 195

#> Downloading 54 of 195

#> Downloading 55 of 195

#> Downloading 56 of 195

#> Downloading 57 of 195

#> Downloading 58 of 195

#> Downloading 59 of 195

#> Downloading 60 of 195

#> Downloading 61 of 195

#> Downloading 62 of 195

#> Downloading 63 of 195

#> Downloading 64 of 195

#> Downloading 65 of 195

#> Downloading 66 of 195

#> Downloading 67 of 195

#> Error in curl::curl_fetch_disk(url, x$path, handle = handle): Unexpected EOF
#> Request failed [ERROR]. Retrying in 1 seconds...

#> Downloading 68 of 195

#> Joining, by = "Identifier"

#> Downloading 69 of 195

#> Downloading 70 of 195

#> Downloading 71 of 195

#> Downloading 72 of 195

#> Downloading 73 of 195

#> Downloading 74 of 195

#> Downloading 75 of 195

#> Downloading 76 of 195

#> Downloading 77 of 195

#> Downloading 78 of 195

#> Downloading 79 of 195

#> Downloading 80 of 195

#> Downloading 81 of 195

#> Joining, by = c("Identifier", "Funding Org:Identifier", "Funding Org:Name", "Funding Org:Department")

#> Downloading 82 of 195

#> Downloading 83 of 195

#> Downloading 84 of 195

#> Downloading 85 of 195

#> Downloading 86 of 195

#> Downloading 87 of 195

#> Downloading 88 of 195

#> Downloading 89 of 195

#> Downloading 90 of 195

#> Downloading 91 of 195

#> Downloading 92 of 195

#> Downloading 93 of 195

#> Downloading 94 of 195

#> New names:
#> * `` -> ...27

#> Downloading 95 of 195

#> Downloading 96 of 195

#> Request failed [404]. Retrying in 1 seconds...

#> Request failed [404]. Retrying in 1 seconds...

#> Request failed: Client error: (404) Not Found

#> Downloading 97 of 195

#> Downloading 98 of 195

#> Downloading 99 of 195

#> Downloading 100 of 195

#> Downloading 101 of 195

#> Downloading 102 of 195

#> Request failed [403]. Retrying in 1 seconds...

#> Request failed [403]. Retrying in 3.9 seconds...

#> Request failed: Client error: (403) Forbidden

#> Downloading 103 of 195

#> Request failed [403]. Retrying in 1 seconds...

#> Request failed [403]. Retrying in 2.7 seconds...

#> Request failed: Client error: (403) Forbidden

#> Downloading 104 of 195

#> Request failed [403]. Retrying in 1.5 seconds...

#> Request failed [403]. Retrying in 1 seconds...

#> Request failed: Client error: (403) Forbidden

#> Downloading 105 of 195

#> Downloading 106 of 195

#> Downloading 107 of 195

#> Downloading 108 of 195

#> Downloading 109 of 195

#> Downloading 110 of 195

#> Downloading 111 of 195

#> Downloading 112 of 195

#> Downloading 113 of 195

#> Downloading 114 of 195

#> Downloading 115 of 195

#> Downloading 116 of 195

#> Downloading 117 of 195

#> Downloading 118 of 195

#> Downloading 119 of 195

#> Downloading 120 of 195

#> Downloading 121 of 195

#> Downloading 122 of 195

#> Downloading 123 of 195

#> Downloading 124 of 195

#> Downloading 125 of 195

#> Downloading 126 of 195

#> Downloading 127 of 195

#> Downloading 128 of 195

#> Downloading 129 of 195

#> Downloading 130 of 195

#> Downloading 131 of 195

#> Downloading 132 of 195

#> Downloading 133 of 195

#> Downloading 134 of 195

#> Downloading 135 of 195

#> Downloading 136 of 195

#> Downloading 137 of 195

#> Downloading 138 of 195

#> Downloading 139 of 195

#> Downloading 140 of 195

#> Downloading 141 of 195

#> Downloading 142 of 195

#> Downloading 143 of 195

#> Downloading 144 of 195

#> Downloading 145 of 195

#> Downloading 146 of 195

#> Downloading 147 of 195

#> Downloading 148 of 195

#> Downloading 149 of 195

#> Downloading 150 of 195

#> Downloading 151 of 195

#> Downloading 152 of 195

#> Downloading 153 of 195

#> Downloading 154 of 195

#> Downloading 155 of 195

#> Downloading 156 of 195

#> Downloading 157 of 195

#> Downloading 158 of 195

#> Downloading 159 of 195

#> Downloading 160 of 195

#> Downloading 161 of 195

#> Downloading 162 of 195

#> Downloading 163 of 195

#> Downloading 164 of 195

#> Downloading 165 of 195

#> Downloading 166 of 195

#> Downloading 167 of 195

#> Downloading 168 of 195

#> Request failed [403]. Retrying in 2 seconds...

#> Request failed [403]. Retrying in 3 seconds...

#> Request failed: Client error: (403) Forbidden

#> Downloading 169 of 195

#> Downloading 170 of 195

#> Downloading 171 of 195

#> Downloading 172 of 195

#> Downloading 173 of 195

#> Downloading 174 of 195

#> Downloading 175 of 195

#> Downloading 176 of 195

#> Downloading 177 of 195

#> Downloading 178 of 195

#> Downloading 179 of 195

#> Downloading 180 of 195

#> Downloading 181 of 195

#> Downloading 182 of 195

#> Downloading 183 of 195

#> Downloading 184 of 195

#> Downloading 185 of 195

#> Downloading 186 of 195

#> Downloading 187 of 195

#> Downloading 188 of 195

#> Downloading 189 of 195

#> Downloading 190 of 195

#> Downloading 191 of 195

#> Downloading 192 of 195

#> Downloading 193 of 195

#> Downloading 194 of 195

#> Downloading 195 of 195

df <- tsg_core_data(grants)

#> Processing 1 of 195

#> Processing 2 of 195

#> Processing 3 of 195

#> Processing 4 of 195

#> Processing 5 of 195

#> Processing 6 of 195

#> Processing 7 of 195

#> Processing 8 of 195

#> Processing 9 of 195

#> Processing 10 of 195

#> Processing 11 of 195

#> Processing 12 of 195

#> Processing 13 of 195

#> Processing 14 of 195

#> Processing 15 of 195

#> Processing 16 of 195

#> Processing 17 of 195

#> Processing 18 of 195

#> Processing 19 of 195

#> Processing 20 of 195

#> Processing 21 of 195

#> Processing 22 of 195

#> Processing 23 of 195

#> Processing 24 of 195

#> Processing 25 of 195

#> Processing 26 of 195

#> Processing 27 of 195

#> Processing 28 of 195

#> Processing 29 of 195

#> Processing 30 of 195

#> Processing 31 of 195

#> Processing 32 of 195

#> Processing 33 of 195

#> Processing 34 of 195

#> Processing 35 of 195

#> Processing 36 of 195

#> Processing 37 of 195

#> Processing 38 of 195

#> Processing 39 of 195

#> Processing 40 of 195

#> Processing 41 of 195

#> Processing 42 of 195

#> Processing 43 of 195

#> Processing 44 of 195

#> Processing 45 of 195

#> Processing 46 of 195

#> Processing 47 of 195

#> Processing 48 of 195

#> Processing 49 of 195

#> Processing 50 of 195

#> Processing 51 of 195

#> Processing 52 of 195

#> Processing 53 of 195

#> Processing 54 of 195

#> Processing 55 of 195

#> Processing 56 of 195

#> Processing 57 of 195

#> Processing 58 of 195

#> Processing 59 of 195

#> Processing 60 of 195

#> Processing 61 of 195

#> Processing 62 of 195

#> Processing 63 of 195

#> Processing 64 of 195

#> Processing 65 of 195

#> Processing 66 of 195

#> Processing 67 of 195

#> Processing 68 of 195

#> Processing 69 of 195

#> Processing 70 of 195

#> Processing 71 of 195

#> Processing 72 of 195

#> Processing 73 of 195

#> Processing 74 of 195

#> Processing 75 of 195

#> Processing 76 of 195

#> Processing 77 of 195

#> Processing 78 of 195

#> Processing 79 of 195

#> Processing 80 of 195

#> Processing 81 of 195

#> Processing 82 of 195

#> Processing 83 of 195

#> Processing 84 of 195

#> Processing 85 of 195

#> Processing 86 of 195

#> Processing 87 of 195

#> Processing 88 of 195

#> Processing 89 of 195

#> Processing 90 of 195

#> Processing 91 of 195

#> Processing 92 of 195

#> Processing 93 of 195

#> Processing 94 of 195

#> Processing 95 of 195

#> Processing 96 of 195

#> Processing 97 of 195

#> Processing 98 of 195

#> Processing 99 of 195

#> Processing 100 of 195

#> Processing 101 of 195

#> Processing 102 of 195

#> Processing 103 of 195

#> Processing 104 of 195

#> Processing 105 of 195

#> Processing 106 of 195

#> Processing 107 of 195

#> Processing 108 of 195

#> Processing 109 of 195

#> Processing 110 of 195

#> Processing 111 of 195

#> Processing 112 of 195

#> Processing 113 of 195

#> Processing 114 of 195

#> Processing 115 of 195

#> Processing 116 of 195

#> Processing 117 of 195

#> Processing 118 of 195

#> Processing 119 of 195

#> Processing 120 of 195

#> Processing 121 of 195

#> Processing 122 of 195

#> Processing 123 of 195

#> Processing 124 of 195

#> Processing 125 of 195

#> Processing 126 of 195

#> Processing 127 of 195

#> Processing 128 of 195

#> Processing 129 of 195

#> Processing 130 of 195

#> Processing 131 of 195

#> Processing 132 of 195

#> Processing 133 of 195

#> Processing 134 of 195

#> Processing 135 of 195

#> Processing 136 of 195

#> Processing 137 of 195

#> Processing 138 of 195

#> Processing 139 of 195

#> Processing 140 of 195

#> Processing 141 of 195

#> Processing 142 of 195

#> Processing 143 of 195

#> Processing 144 of 195

#> Processing 145 of 195

#> Processing 146 of 195

#> Processing 147 of 195

#> Processing 148 of 195

#> Processing 149 of 195

#> Processing 150 of 195

#> Processing 151 of 195

#> Processing 152 of 195

#> Processing 153 of 195

#> Processing 154 of 195

#> Processing 155 of 195

#> Processing 156 of 195

#> Processing 157 of 195

#> Processing 158 of 195

#> Processing 159 of 195

#> Processing 160 of 195

#> Processing 161 of 195

#> Processing 162 of 195

#> Processing 163 of 195

#> Processing 164 of 195

#> Processing 165 of 195

#> Processing 166 of 195

#> Processing 167 of 195

#> Processing 168 of 195

#> Processing 169 of 195

#> Processing 170 of 195

#> Processing 171 of 195

#> Processing 172 of 195

#> Processing 173 of 195

#> Processing 174 of 195

#> Processing 175 of 195

#> Processing 176 of 195

#> Processing 177 of 195

#> Processing 178 of 195

#> Processing 179 of 195

#> Processing 180 of 195

#> Processing 181 of 195

#> Processing 182 of 195

#> Processing 183 of 195

#> Processing 184 of 195

#> Processing 185 of 195

#> Processing 186 of 195

#> Processing 187 of 195

#> Processing 188 of 195

#> Processing 189 of 195

#> Processing 190 of 195

#> Processing 191 of 195

#> Processing 192 of 195

#> Processing 193 of 195

#> Processing 194 of 195

#> Processing 195 of 195

# }